[#37] Acknowledge Server ID in start
Currently the server ID received in the start of MsgPythonShell is not
acknowledged in start which may trigger the server to resend the server
ID reply.
