+++ tests/shells/conftest.py
@@ -1,6 +1,18 @@
+from contextlib import contextmanager
+from collections import namedtuple
+import mock
 import pytest
 from crl.interactivesessions.shells.bashshell import BashShell
 from crl.interactivesessions.shells.shell import DEFAULT_STATUS_TIMEOUT
+from crl.interactivesessions.shells.msgpythonshell import MsgPythonShell
+from crl.interactivesessions.shells.terminalclient import TerminalClient
+from crl.interactivesessions.shells.remotemodules.msgmanager import (
+    StrComm,
+    Retry)
+
+from .loststrcomm import (
+    LostStrComm,
+    CustomTerminalComm)
 from .bashterminalshell import BashTerminalShell
 from .statuscodeverifier import StatusCodeVerifier
 
@@ -51,3 +63,142 @@ def statuscodeverifier(bash_shell_with_terminal, bash_terminal, status_timeout):
     return StatusCodeVerifier(bash_shell_with_terminal=bash_shell_with_terminal,
                               bash_terminal=bash_terminal,
                               timeout=status_timeout)
+
+
+@pytest.fixture
+def msgpythonshell(normal_pythonterminal):
+    with msgpythonshell_context(normal_pythonterminal) as m:
+        yield m
+
+
+@pytest.fixture
+def shortretry_msgpythonshell(retry_shellcontext):
+    with retry_shellcontext(Retry(tries=20, interval=0.4, timeout=0.5)) as s:
+        yield s
+
+
+@pytest.fixture
+def terminate_msgpythonshell(retry_shellcontext, normal_pythonterminal):
+    with retry_shellcontext(Retry(tries=3, interval=0.4, timeout=0.5)) as s:
+        try:
+            yield s
+        finally:
+            normal_pythonterminal.terminate()
+
+# pylint: disable=unused-argument
+@pytest.fixture
+def retry_shellcontext(mock_strcomm, simple_retry_shellcontext):
+    return simple_retry_shellcontext
+
+
+@pytest.fixture
+def simple_retry_shellcontext(normal_pythonterminal):
+    @contextmanager
+    def ctx(retry):
+        try:
+            MsgPythonShell.set_retry(retry)
+            with msgpythonshell_context(normal_pythonterminal) as m:
+                yield m
+        finally:
+            MsgPythonShell.reset_retry()
+
+    return ctx
+
+
+@pytest.fixture
+def client_rubbish_msgpythonshell(client_rubbish_pythonterminal_ctx):
+    with msgpythonshell_context(client_rubbish_pythonterminal_ctx.pythonterminal) as m:
+        yield m
+
+
+@pytest.fixture
+def server_rubbish_msgpythonshell(server_rubbish_pythonterminal_ctx):
+    with msgpythonshell_context(server_rubbish_pythonterminal_ctx.pythonterminal) as m:
+        yield m
+
+
+@pytest.fixture
+def mock_strcomm(request):
+    lst = LostStrComm(**request.param)
+
+    def strcomm_fact(*args, **kwargs):
+        s = StrComm(*args, **kwargs)
+        lst.set_strcomm(s)
+        return lst
+
+    with customterminalcomm_context():
+        with mock.patch('crl.interactivesessions.shells.'
+                        'remotemodules.msgmanager.StrComm') as p:
+            p.side_effect = strcomm_fact
+            yield lst
+
+
+@contextmanager
+def customterminalcomm_context():
+    with mock.patch('crl.interactivesessions.shells.'
+                    'terminalclient.TerminalComm',
+                    side_effect=CustomTerminalComm) as p:
+        yield p
+
+
+class ShellContext(namedtuple('ShellContext', ['shell', 'context'])):
+    pass
+
+
+@pytest.fixture
+def client_rubbish_pythonterminal_ctx(rubbish_context, request):
+    with rubbish_context['client'] as ctx:
+        yield ctx
+
+
+@pytest.fixture
+def server_rubbish_pythonterminal_ctx(rubbish_context, request):
+    with rubbish_context['server'] as ctx:
+        yield ctx
+
+
+@pytest.fixture
+def chunk_msgpythonshell(chunk_pythonterminal_ctx):
+    with msgpythonshell_context(chunk_pythonterminal_ctx.pythonterminal) as m:
+        yield m
+
+
+class CustomTerminalClient(TerminalClient):
+    def __init__(self):
+        super(CustomTerminalClient, self).__init__()
+        self.received_msgs = []
+
+    def _receive(self, timeout):
+        msg = super(CustomTerminalClient, self)._receive(timeout)
+        self.received_msgs.append(msg)
+        return msg
+
+
+@pytest.fixture
+def customterminalclient():
+    c = CustomTerminalClient()
+    with mock.patch('crl.interactivesessions.shells.msgpythonshell.TerminalClient',
+                    spec_set=True) as p:
+        p.return_value = c
+        yield c
+
+
+@pytest.fixture
+def custommsgpythonshell(customterminalclient, simple_retry_shellcontext):
+    with simple_retry_shellcontext(Retry(tries=50, interval=0.01, timeout=0.5)) as s:
+        yield s
+
+
+@contextmanager
+def msgpythonshell_context(terminal):
+    m = MsgPythonShell()
+    m.set_terminal(terminal)
+    m.delaybeforesend = 1
+    m.start()
+    assert m.delaybeforesend == 0
+    try:
+        yield m
+    finally:
+        m.exit()
+        assert m.delaybeforesend == 1
+        terminal.join(timeout=3)
